Refactor course leaderboard entries handler into separate file
